requirements.txt
streamlit
langgraph
langchain
langchain-aws
langchain-core
pydantic
python-dotenv
boto3
reportlab